% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{c.spec_fam}
\alias{c.spec_fam}
\title{S3 method for generic c() for \code{\linkS4class{spec_fam}} objects.}
\usage{
\method{c}{spec_fam}(...)
}
\description{
S3 method for generic c() for \code{\linkS4class{spec_fam}} objects.
}
