% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R, R/families.R
\name{confidence}
\alias{confidence}
\alias{confidence.mmi_lm}
\alias{confidence.mmi_logreg}
\alias{confidence.mmi_lmm}
\alias{confidence.fam}
\title{Generic method to compute confidence regions for mmi model and family objects.}
\usage{
confidence(object, level, ...)

\method{confidence}{mmi_lm}(object, level)

\method{confidence}{mmi_logreg}(object, level)

\method{confidence}{mmi_lmm}(object, level)

\method{confidence}{fam}(object, level = 0.95)
}
\arguments{
\item{object}{Either a mmi model object or a mmi family object.}

\item{level}{The confidence level.}
}
\description{
For \code{\linkS4class{mmi_lm}} objects a confidence interval is constructed using the
t-distribution assumption for the linear parameter over the standard error. The standard
error is estimated using sandwich estimation. For the other models the confidence intervals
are computed using the profile likelihood. A tidy tibble will be setup with the information.
For a \code{\linkS4class{fam}} object confidence intervals will be computed for each model
element using the individual methods, and the tibbles will then be binded together in a
map-reduce pattern.
}
\section{Methods (by class)}{
\itemize{
\item \code{mmi_lm}: Computes the confidence interval using the t-distribution assumption and sets up a
tidy tibble with the information. Uses the sandwich estimated standard errors.

\item \code{mmi_logreg}: Computes the confidence interval using profile likelihood (the confint method)
and sets up a tidy tibble with the information.

\item \code{mmi_lmm}: Computes the confidence interval using profile likelihood (the confint method)
and sets up a tidy tibble with the information.

\item \code{fam}: Maps over the family, constructs a confidence interval for each model object using the
corresponding S3 method, and the binds the information together in a tidy way.
}}

