% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecrf}
\alias{ecrf}
\title{Demographical and clinical data from the article  "Natural variation in immune cell parameters is preferentially driven by genetic factors".}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 816 rows and 44 columns.}
\usage{
ecrf
}
\description{
A dataset containing 45 clinical and demographical variables from From the Institut Pasteur project Milieu Interieur.
This data was analyzed in the article "Natural variation in immune cell parameters is preferentially driven by genetic factors"
using the mmi R package.
}
\keyword{datasets}
