% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facs}
\alias{facs}
\title{Flow cytometry data from the article  "Natural variation in immune cell parameters is preferentially driven by genetic factors".}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 816 rows and 167 columns.}
\usage{
facs
}
\description{
A dataset containing 167 immunophenotypes. From the Institut Pasteur project Milieu Interieur.
This data was analyzed in the article "Natural variation in immune cell parameters is preferentially driven by genetic factors"
using the mmi R package.
}
\keyword{datasets}
