% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facs_annotation}
\alias{facs_annotation}
\title{Names and annotation for the immunophenotypes in the facs table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 166 rows and 3 columns.}
\usage{
facs_annotation
}
\description{
Names and annotation for the immunophenotypes in the facs table
}
\keyword{datasets}
