% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{class}
\name{fam-class}
\alias{fam-class}
\alias{.make_fam}
\title{S4 class "fam"}
\description{
List of family objects.
}
\details{
Class that represents a list of model objects as a vehicle for looping and
then bind together information from the model objects, basically a map-reduce
pattern. The object also defines a family in the multiple testing sense.
}
