% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.spec_lm}
\alias{fit_model.spec_logreg}
\alias{fit_model.spec_lmm}
\alias{fit_model.spec_trans_lm}
\alias{fit_model.spec_trans_lmm}
\alias{fit_model.spec_beta}
\title{S3 Generic function fit_model to construct mmi_model objects.}
\usage{
fit_model(object, study_frame)

\method{fit_model}{spec_lm}(object, study_frame)

\method{fit_model}{spec_logreg}(object, study_frame)

\method{fit_model}{spec_lmm}(object, study_frame)

\method{fit_model}{spec_trans_lm}(object, study_frame)

\method{fit_model}{spec_trans_lmm}(object, study_frame)

\method{fit_model}{spec_beta}(object, study_frame)
}
\arguments{
\item{object}{\code{\linkS4class{spec}} object.}

\item{study_frame}{The data frame that includes all variables in all models}
}
\description{
Sets up the formula using \code{\link{get_formula}} and then fits an
an appropriate model using the data in \code{study_frame} and adds to the specifiation object
creating a mmi model object.
}
\section{Methods (by class)}{
\itemize{
\item \code{spec_lm}: Constructs a \code{\linkS4class{mmi_lm}} object.

Sets up the formula using \code{\link{get_formula}} and then fits an
\code{\link[stats]{lm}} using the formula and the data in study_frame.

\item \code{spec_logreg}: Constructs a \code{\linkS4class{mmi_logreg}} object by setting up a formula using \code{\link{get_formula}}
and then fitting a \code{\link[stats]{glm}} logistic regression using the formula and the data in study_frame.

\item \code{spec_lmm}: Constructs a \code{\linkS4class{mmi_lmm}} object by setting up a formula using \code{\link{get_formula}}
and then fitting a \code{\linkS4class{lmerMod}} using the formula and the data in study_frame.

\item \code{spec_trans_lm}: Constructs a \code{\linkS4class{mmi_trans_lm}} object by setting up a formula using \code{\link{get_formula}}
and then fitting a \code{\link[stats]{lm}} on the transformed response using the formula and the data in study_frame.

\item \code{spec_trans_lmm}: Constructs a \code{\linkS4class{mmi_trans_lmm}} object by setting up a formula using \code{\link{get_formula}}
and then fitting a \code{\linkS4class{lmerMod}} on the transformed response using the formula and the data in study_frame.

\item \code{spec_beta}: Constructs a \code{\linkS4class{mmi_logreg}} object by setting up a formula using \code{\link{get_formula}}
and then fitting a \code{\link[stats]{glm}} logistic regression using the formula and the data in study_frame.
}}

