% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/families.R
\name{ft}
\alias{ft}
\alias{ft.mmi_lm}
\alias{ft.mmi_lmm}
\alias{ft.fam}
\title{Generic method for F-tests.}
\usage{
ft(object)

\method{ft}{mmi_lm}(object)

\method{ft}{mmi_lmm}(object)

\method{ft}{fam}(object)
}
\arguments{
\item{object}{Either an mmi model object or an mmi model family object.}
}
\description{
A method is implemented for all the mmi model classes. A method is also implemented for
the family class. In that case the method is basically a loop over all models stored in
the family and then the method for the model is called for each model element. If the method
is called on a model object a tidy tibble will be set up with the result for the treatment
or the treatment levels. If the method is called on a family object individual tibbles will
be set up for each model element which will then be binded together in a map-reduce pattern.
}
\details{
Will loop through all treatment variables and fit null models for them. This is done using
update methods for the model fit if that works (depends on how the model fit is implemented),
otherwise it's done by using update on the formula. P values are
then computed using appropriate functions for each model class.
For \code{\linkS4class{mmi_lmm}} objects
\code{\link[pbkrtest]{KRmodcomp}} Kenward-Rogers F-test are used.
}
\section{Methods (by class)}{
\itemize{
\item \code{mmi_lm}: Performs a F test for the \code{\link[stats]{lm}} object in the
slot fit and sets up a tibble for the result.

\item \code{mmi_lmm}: Performs a F test for the \code{\link[lme4]{lmerMod}} object in the slot fit
using a \code{\link[pbkrtest]{KRmodcomp}} Kenward-Rogers F-test
and sets up a tibble with the result.

\item \code{fam}: Loops over all model elements in the family and does a ft test for each, if defined;
tidy tibbles with this information will then be binded together.
Also computes the FDR-adjusted significance level.
}}

