% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{get_formula}
\alias{get_formula}
\title{Generic method that computes the formula for the fit.}
\usage{
get_formula(object)
}
\arguments{
\item{object}{A \code{\linkS4class{spec}} object.}
}
\description{
Methods for this generic are implemented for the different \code{\linkS4class{spec}}
classes.
}
