% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{get_null_formula}
\alias{get_null_formula}
\title{Generic method that computes the formula without the treatment for the fit.}
\usage{
get_null_formula(object)
}
\arguments{
\item{object}{A \code{\linkS4class{spec}} object.}
}
\description{
Methods for this generic are implemented for the different \code{\linkS4class{spec}}
classes.
}
