% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{get_trt_levels}
\alias{get_trt_levels}
\title{Generic method that computes the names given to effects from variables that are
factors/characters/logicals.}
\usage{
get_trt_levels(object, fit)
}
\arguments{
\item{object}{A \code{\linkS4class{spec}} object.}

\item{fit}{The model fit that will be added to the \code{\linkS4class{spec}} in the subsequent
mmi model construction.}
}
\description{
Methods for this generic are implemented for the different \code{\linkS4class{spec}}
classes.
}
