% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{methods}
\name{initialize,spec-method}
\alias{initialize,spec-method}
\title{S4 initialize method for \code{\linkS4class{spec}}.}
\usage{
\S4method{initialize}{spec}(.Object, ...)
}
\description{
S4 initialize method for \code{\linkS4class{spec}}.
}
