% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{methods}
\name{initialize,spec_logreg-method}
\alias{initialize,spec_logreg-method}
\title{S4 initialize method for \code{\linkS4class{spec_logreg}}.}
\usage{
\S4method{initialize}{spec_logreg}(.Object, ...)
}
\arguments{
\item{.Object}{The \code{\linkS4class{spec}} object to be created.}

\item{...}{Arguments from the constructor.}
}
\description{
Makes sure that the transformation for the logistic regression class is a log.
}
