% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{methods}
\name{initialize,spec_nb-method}
\alias{initialize,spec_nb-method}
\title{S4 initialize method for \code{\linkS4class{spec_nb}}.}
\usage{
\S4method{initialize}{spec_nb}(.Object, ...)
}
\arguments{
\item{.Object}{The \code{\linkS4class{spec}} object to be created.}

\item{...}{Arguments from the constructor.}
}
\description{
S4 initialize method for \code{\linkS4class{spec_nb}}.
}
