% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{methods}
\name{initialize,spec_trans_lm-method}
\alias{initialize,spec_trans_lm-method}
\title{S4 initialize method for \code{\linkS4class{spec_trans_lm}}.}
\usage{
\S4method{initialize}{spec_trans_lm}(.Object, ...)
}
\arguments{
\item{.Object}{The \code{\linkS4class{spec}} object to be created.}

\item{...}{Arguments from the constructor.}
}
\description{
Makes sure that the lm transformation class has a defined transformation
}
