% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{lrt.mmi_lm}
\alias{lrt.mmi_lm}
\title{Performs a likelihood ratio test (an F-test) for the \code{\link[stats]{lm}} object in the
slot fit and sets up a tidy tibble for the result.}
\usage{
\method{lrt}{mmi_lm}(object)
}
\description{
Performs a likelihood ratio test (an F-test) for the \code{\link[stats]{lm}} object in the
slot fit and sets up a tidy tibble for the result.
}
