% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{lrt.mmi_lmm}
\alias{lrt.mmi_lmm}
\title{Performs a likelihood ratio test for the \linkS4class{lmerMod} in the fit slot and sets up
a tidy tibble for the result.}
\usage{
\method{lrt}{mmi_lmm}(object)
}
\description{
Performs a likelihood ratio test for the \linkS4class{lmerMod} in the fit slot and sets up
a tidy tibble for the result.
}
