% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lrt.mmi_logreg}
\alias{lrt.mmi_logreg}
\title{Performs a likelihood ratio test for the binomial \code{\link[stats]{glm}} in the fit slot
and sets up a tidy tibble with the result.}
\usage{
\method{lrt}{mmi_logreg}(object)
}
\description{
Performs a likelihood ratio test for the binomial \code{\link[stats]{glm}} in the fit slot
and sets up a tidy tibble with the result.
}
