% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{make_fam}
\alias{make_fam}
\title{Constructs fam objects from a specification family.}
\usage{
make_fam(spec_fam, study_frame)
}
\arguments{
\item{spec_fam}{A specification family object.}

\item{study_frame}{The data to be investigated. Must include all response, treatment and
control variables in the specification family.}
}
\description{
Main function to create fam objects, lists of model objects. When make_fam is called
models will be fit to the data in \code{study_frame} according to each specification.
}
