% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{mmi_beta-class}
\alias{mmi_beta-class}
\alias{.make_beta}
\title{S4 class for representing a beta regression specification and a fit to that specification.}
\description{
Inherits from \code{\linkS4class{spec_beta}} and \code{\linkS4class{mmi_model}}.
A \code{\link[betareg]{betareg}} logit-lik fit is added to the specification object.
}
\section{Slots}{

\describe{
\item{\code{fit}}{The fitted \code{\link[betareg]{betareg}} object.}
}}

