% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{mmi_lmm-class}
\alias{mmi_lmm-class}
\alias{.make_lmm}
\title{S4 class for representing a linear mixed model specification and a fit to that specification.}
\description{
Inherits from \code{\linkS4class{spec_lmm}} and \code{\linkS4class{mmi_model}}.
Contains the \code{\linkS4class{lmerMod}} that us fitted according to the specification.
}
\section{Slots}{

\describe{
\item{\code{fit}}{The \code{\linkS4class{lmerMod}} fit to the contained model specification.}
}}

