% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{mmi_logreg-class}
\alias{mmi_logreg-class}
\alias{.make_logreg}
\title{S4 class for representing a logistic regression specification and a fit to that specification.}
\description{
Inherits from \code{\linkS4class{spec_logreg}} and \code{\linkS4class{mmi_model}}.
A \code{\link[stats]{glm}} binomial logit-link fit is added to the specification object.
}
\section{Slots}{

\describe{
\item{\code{fit}}{The fitted \code{\link[stats]{glm}} object.}
}}

