% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{mmi_model-class}
\alias{mmi_model-class}
\alias{.make_model}
\title{S4 class for a model representation and model fit. This is the superclass which defines
the parsed variable terms in the model matrix.}
\description{
S4 class for a model representation and model fit. This is the superclass which defines
the parsed variable terms in the model matrix.
}
\section{Slots}{

\describe{
\item{\code{var_labels}}{Parsed variable and variable level names that appears in model.matrix}
}}

