% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{mmi_nb-class}
\alias{mmi_nb-class}
\alias{.make_nb}
\title{S4 class for representing a negative binomial regression specification and a fit to that specification.}
\description{
Inherits from \code{\linkS4class{spec_nb}} and \code{\linkS4class{mmi_model}}.
A \code{\link[MASS]{glm.nb}} negative binomial fit is added to the specification object.
}
\section{Slots}{

\describe{
\item{\code{fit}}{The fitted \code{\link[MASS]{glm.nb}} object.}
}}

