% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.spec_fam}
\alias{print.spec_fam}
\title{S3 method for generic print for spec_fam objects.}
\usage{
\method{print}{spec_fam}(x, ...)
}
\arguments{
\item{x}{The \code{\linkS4class{spec_fam}} object to be printed.}

\item{...}{Not used currently.}
}
\description{
S3 method for generic print for spec_fam objects.
}
