% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/families.R
\name{prop_var}
\alias{prop_var}
\alias{prop_var.fam}
\title{Extract estimate of variance component to the treatment random effects}
\usage{
prop_var(object)

\method{prop_var}{fam}(object)
}
\description{
Extract estimate of variance component to the treatment random effects
}
\section{Methods (by class)}{
\itemize{
\item \code{fam}: Loops over all model elements in the family and estimates the variance component for each
treatment random effect;
tidy tibbles with this information will then be binded together.
}}

