% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{select_confidence}
\alias{select_confidence}
\title{Calculates FCR adjusted confidence intervals}
\usage{
select_confidence(fam, test_tib, crit = "FDR", thresh = 0.05,
  level = 0.95)
}
\arguments{
\item{fam}{A \code{\linkS4class{fam}} object where models will be selected from.}

\item{test_tib}{A tibble with results from \code{\link{test}} or \code{\link{anova}} from
the fam object supplied with the first argument.}

\item{crit}{The criterion used for selection. Has to be a significance column in test_tib.}

\item{thresh}{The threshold for the criterion: select all parameters who had higher significance than this.}

\item{level}{The FCR confidence level.}
}
\description{
Takes a \code{\linkS4class{fam}} and selects which of the fits to calculate confidence
intervals for based on the significance values of tests given in the tibble test_tib which
typically is an output of \code{\link{test}} or \code{\link{anova}}. If a response-treatment
pair is significant based on the crit and thresh parameters then confidence levels willl be
calculated for all treatment level parameters for that pair. The confidence intervals are
FCR-adjusted based on the number of parameters selected in total.
}
