% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{methods}
\name{show,fam-method}
\alias{show,fam-method}
\title{S4 method for generic show for fam objects}
\usage{
\S4method{show}{fam}(object)
}
\arguments{
\item{object}{The \code{\linkS4class{fam}} object to be shown.}
}
\description{
S4 method for generic show for fam objects
}
