% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{methods}
\name{show,spec_fam-method}
\alias{show,spec_fam-method}
\title{S4 method for generic show for spec_fam objects}
\usage{
\S4method{show}{spec_fam}(object)
}
\arguments{
\item{object}{The \code{\linkS4class{spec_fam}} object to be shown.}
}
\description{
S4 method for generic show for spec_fam objects
}
