% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{class}
\name{spec-class}
\alias{spec-class}
\alias{.make_spec}
\title{S4 class spec that represents a model specification.}
\description{
Contains a specification of variables that should be included for the fitting of one model.
Has 4 subclasses that has the same specification and also defines what model to use.
Currently the subclasses \code{\linkS4class{spec_lm}}, \code{\linkS4class{spec_lm}},
\code{\linkS4class{spec_logreg}} and \code{\linkS4class{spec_trans_lm}} are implemented for
linear models, linear mixed models, logistic regression models and
transformed linear models respectively.
}
\section{Slots}{

\describe{
\item{\code{response}}{Response variable.}

\item{\code{treatment}}{Treatment variable.}

\item{\code{controls}}{Control variables.}

\item{\code{trans}}{If the response is transformed in any way.
Will be used to back-transform estimates and confidence intervals.}
}}

