% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{class}
\name{spec_beta-class}
\alias{spec_beta-class}
\alias{.make_spec_beta}
\title{S4 class spec_nb for a beta regression specification.}
\description{
See help for \code{\linkS4class{spec}}.
}
