% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{class}
\name{spec_lmm-class}
\alias{spec_lmm-class}
\alias{.make_spec_lmm}
\title{S4 class spec_lmm for a linear model with random effects specification.}
\description{
See help for \code{\linkS4class{spec}}.
}
