% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{class}
\name{spec_logreg-class}
\alias{spec_logreg-class}
\alias{.make_spec_logreg}
\title{S4 class spec_logreg for a logistic regression specification.}
\description{
See help for \code{\linkS4class{spec}}.
}
