% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\docType{class}
\name{spec_trans_lm-class}
\alias{spec_trans_lm-class}
\alias{.make_spec_trans_lm}
\title{S4 class "spec_trans_lm_lm" for linear model with a transformed response.}
\description{
Using this specification will transformed the response variable.
For the slots see help for \code{\linkS4class{spec}}.
}
