% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{specify}
\alias{specify}
\title{Constructs a list of model specifications from variable names.}
\usage{
specify(responses, treatments = character(0), controls = character(0),
  model = NULL, trans = NULL, rands = NULL)
}
\arguments{
\item{responses}{Character vector of response variables.}

\item{treatments}{Character vector of treatment variables.}

\item{controls}{Character vector of control variables that will be shared for
all models.}

\item{model}{Which model should be fitted?}

\item{trans}{Are the response variables transformed?}

\item{rands}{Possible random effects shared by the response variables.}
}
\description{
Every combination of response and treatment will be a \code{spec} object in
the list. They will all share the properties in \code{controls}, \code{model},
\code{trans} and (if the model is a mixed model) in \code{rands}. So all
models will have the same control variables for instance.
}
