% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{specify_with_list}
\alias{specify_with_list}
\title{Creates a \code{\linkS4class{spec_fam}} from a list where each element is a list with
elements response, treatment and control which all carries a string formula, and model
which states the distributional assumption, i.e. the model to fit.}
\usage{
specify_with_list(spec_list)
}
\description{
Creates a \code{\linkS4class{spec_fam}} from a list where each element is a list with
elements response, treatment and control which all carries a string formula, and model
which states the distributional assumption, i.e. the model to fit.
}
