% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{[.fam}
\alias{[.fam}
\title{S3 method for generic `[` for \code{\linkS4class{fam}} objects.}
\usage{
\method{[}{fam}(x, i, ...)
}
\description{
S3 method for generic `[` for \code{\linkS4class{fam}} objects.
}
