% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{[.spec_fam}
\alias{[.spec_fam}
\title{S3 method for generic `[` for \code{\linkS4class{spec_fam}} objects.}
\usage{
\method{[}{spec_fam}(x, idx, ...)
}
\description{
S3 method for generic `[` for \code{\linkS4class{spec_fam}} objects.
}
