% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/families.R
\name{wald}
\alias{wald}
\alias{wald.mmi_lm}
\alias{wald.fam}
\title{A method is implemented for all the mmi model classes. A method is also implemented for
the family class. In that case the method is basically a loop over all models stored in
the family and then the method for the model is called for each model element. If the method
is called on a model object a tidy tibble will be set up with the result for the treatment
or the treatment levels. If the method is called on a family object individual tibbles will
be set up for each model element which will then be binded together in a map-reduce pattern.}
\usage{
wald(object)

\method{wald}{mmi_lm}(object)

\method{wald}{fam}(object)
}
\arguments{
\item{object}{Either an mmi model object or an mmi model family object.}
}
\description{
Currently only implemented for \code{\linkS4class{mmi_lm}} objects. Standard wald test.
}
\section{Methods (by class)}{
\itemize{
\item \code{mmi_lm}: Performs a wald test for \code{\linkS4class{mmi_lm}} objects.

\item \code{fam}: Loops over all model elements in the family and does a Wald test for each, if defined;
tidy tibbles with this information will then be binded together.
Also computes the FDR-adjusted significance level.
}}

