% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{class}
\name{fam-class}
\alias{fam-class}
\alias{.make_fam}
\title{S4 class "fam"}
\description{
List of family objects.
}
\details{
Class that represents a complete analysis. It's a list of model objects
create to be a vehicle for looping and
then binding together information from the model objects in a map-reduce like
pattern. The object also defines a family in the multiple testing sense.
}
