% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/families.R
\name{lrt}
\alias{lrt}
\alias{lrt.mmi_model}
\alias{lrt.fam}
\title{Generic method for likelihood ratio tests.}
\usage{
lrt(object)

\method{lrt}{mmi_model}(object)

\method{lrt}{fam}(object)
}
\arguments{
\item{object}{Either an mmi model object or an mmi model family object.}
}
\description{
A method is implemented for all the mmi model classes. A method is also implemented for
the family class. In that case the method is basically a loop over all models stored in
the family and then the method for the model is called for each model element. If the method
is called on a model object a tidy tibble will be set up with the result for the treatment
or the treatment levels. If the method is called on a family object individual tibbles will
be set up for each model element which will then be binded together in a map-reduce pattern.
}
\details{
Will loop through all treatment variables and fit null models for them. This is done using
update methods for the model fit if that works (depends on how the model fit is implemented),
otherwise it's done by using update on the \code{\link{get_formula}} formula. P values are
then computed using the \code{\link{logLik}} method on the null and full models.
}
\section{Methods (by class)}{
\itemize{
\item \code{mmi_model}: Performs a likelihood ratio test for the \code{\link[stats]{lm}} object in the
slot fit and sets up a tibble for the result.

\item \code{fam}: Loops over all model elements in the family and does a lrt test for each; tidy tibbles with
this information will then be binded together. Also computes the FDR-adjusted significance
level.
}}

