% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{class}
\name{spec_fam-class}
\alias{spec_fam-class}
\alias{.make_spec_fam}
\title{An S4 class representing a family of specifications}
\description{
Basically a list of specification objects. Can hold any of the specification objects.
It's fate is to become to a fam object using make_fam(obj).
}
\section{Slots}{

\describe{
\item{\code{responses}}{All response variables for this family.}

\item{\code{treatments}}{All treatment variables for this family.}
}}

