% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{specify}
\alias{specify}
\title{Constructs a list of model specifications from formula strings.}
\usage{
specify(responses, treatments = NULL, controls = NULL, model,
  trans = NULL)
}
\arguments{
\item{responses}{Character vector of response formula strings.}

\item{treatments}{Character vector of treatment formula strings.}

\item{controls}{Control formula string that will be shared for all models.}

\item{model}{Distributional assumption.}

\item{trans}{The pretransformation of response variables.}
}
\description{
Main function to create analysis representations. \code{responses}, \code{treatments}
and \code{controls} should be vectors of formula strings.
Every combination of \code{response} and \code{treatment} formulas
will be a \code{spec} object in the list.
They will all share the properties in \code{controls} and \code{model}.
\code{trans} should be given if its known that a response variable in the database is
transformed and a backtransformation is wanted.
}
